scripts/bot/requirements.txt
python-dotenv
tweepy
gpt_2_simple
tqdm
fire
tensorflow==1.15